2005-10-19  Raphael Poss  <raph@lrde.epita.fr>
       * include/vaucanson/algebra/implementation/series/rat/star_height_visitor.hxx (StarHeighVisitor::sum_or_product):
       Change the type of "post" and "left_height" to "size_t" for
       compatibility with class attribute "h_". Needed for G++ 4 which
       complains about std::max not being defined for (size_t, unsigned).

+++ include/vaucanson/algebra/implementation/series/rat/star_height_visitor.hxx
@@ -30,9 +30,9 @@ namespace vcsn {
     StarHeightVisitor<M_,W_>::sum_or_product(const Node<M_, W_>* left_,
 					     const Node<M_, W_>* right_)
     {
-      unsigned post = h_;
+      size_t post = h_;
       left_->accept(*this);
-      unsigned left_height = h_;
+      size_t left_height = h_;
       h_ = post;
       right_->accept(*this);
       h_ = std::max(h_, left_height);
